requirements.txt
openml
numpy
pandas
simplejson
matplotlib
umap-learn
pymfe
